comprinno_pr_agent/requirements.txt
boto3==1.40.45
PyYAML==6.0.1
Jinja2==3.1.6
esprima==4.0.1
javalang==0.13.0
radon==6.0.1
PyGithub==2.1.1
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
